(video filters) Fix undefined shifts in epx and ntsc
Found by running every filter in gfx/video_filters through the real
softfilter entry points under UBSan.

epx_generic_rgb565 assembles its output pairs as (colorX << 16) + colorY
with colorX held in a uint16_t. That promotes to int, so any pixel from
0x8000 up shifts into the sign bit and the shift is undefined - eight
sites, and UBSan trips on ordinary content ("left shift of 39808 by 16
places cannot be represented in type 'int'"). The five temporaries are
now uint32_t, which is what the uint32_t* destination wanted anyway.
The arithmetic becomes unsigned and the stored bits are unchanged.

ntsc's notch filter shifts signed chroma left: cbuf holds values either
side of zero and (cbuf[x] << 1) / (cbuf[i_m1] << 2) are undefined when
they are negative. Written as * 2 and * 4, which is defined, and every
compiler emits the same instruction.

Both verified byte-identical to the previous filters through the real
softfilter API, comparing dlopen'd builds of each version: 60 cases for
epx and 120 for ntsc, spanning widths 160-512 against heights 144-240,
1/2/4 threads and both pixel formats, four frames each.

Builds clean under gcc x86-64, aarch64-linux-gnu, arm-linux-gnueabihf
and mingw-w64. Zero UBSan reports remain for either filter.

+++ gfx/video_filters/epx.c
@@ -107,7 +107,7 @@ static void epx_generic_rgb565 (unsigned width, unsigned height,
       int first, int lsat, uint16_t *src,
       unsigned src_stride, uint16_t *dst, unsigned dst_stride)
 {
-   uint16_t colorA;
+   uint32_t colorA;
    int w;
 
    for (; height; height--)
@@ -119,10 +119,10 @@ static void epx_generic_rgb565 (unsigned width, unsigned height,
       uint32_t *dP2   = (uint32_t *) (dst + dst_stride);
 
       /* left edge */
-      uint16_t colorX = *sP;
-      uint16_t colorC = *++sP;
-      uint16_t colorB = *lP++;
-      uint16_t colorD = *uP++;
+      uint32_t colorX = *sP;
+      uint32_t colorC = *++sP;
+      uint32_t colorB = *lP++;
+      uint32_t colorD = *uP++;
 
       if ((colorX != colorC) && (colorB != colorD))
       {
